backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
redis==5.2.0
python-multipart==0.0.12
aiofiles==24.1.0
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0
